+++ series_api/rating/models.py
@@ -6,4 +6,5 @@ from series.models import Serie
 class Rating(models.Model):
   serie_id = models.ForeignKey(Serie, on_delete=models.CASCADE)
   votes = models.IntegerField()
-  user_rating = models.DecimalField(max_digits=2, decimal_places=2)
+  # user_rating = models.DecimalField(max_digits=2, decimal_places=2)
+  user_rating = models.FloatField()
